+++ src/Microsoft.Azure.SignalR/DependencyInjectionExtensions.cs
@@ -2,13 +2,9 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Routing;
 using Microsoft.AspNetCore.SignalR;
 using Microsoft.Azure.SignalR;
 using Microsoft.Azure.SignalR.Protocol;
-using Microsoft.Azure.SignalR.Startup;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Options;
@@ -31,23 +27,7 @@ namespace Microsoft.Extensions.DependencyInjection
         /// </remarks>
         public static ISignalRServerBuilder AddAzureSignalR(this ISignalRServerBuilder builder)
         {
-
             builder.Services.AddSingleton<IConfigureOptions<ServiceOptions>, ServiceOptionsSetup>();
-
-#if NETCOREAPP3_0
-            // Load Azure:SignalR:Enabled to determine whether fallback to local SignalR
-            var config = new ConfigurationBuilder()
-                .SetBasePath(Environment.CurrentDirectory)
-                .AddJsonFile("appsettings.json", optional: false)
-                .Build();
-            var isEnabled = config.GetSection(Constants.AzureSignalREnabledKey).Value;
-            // Only apparently false will turn-off Azure SignalR Service
-            if (isEnabled != null && isEnabled.Equals("false", StringComparison.OrdinalIgnoreCase))
-            {
-                return builder;
-            }
-#endif
-
             return builder.AddAzureSignalRCore();
         }
 
@@ -95,13 +75,6 @@ namespace Microsoft.Extensions.DependencyInjection
 
             // If a custom router is added, do not add the default router
             builder.Services.TryAddSingleton(typeof(IEndpointRouter), typeof(DefaultEndpointRouter));
-
-#if NETCOREAPP3_0
-            builder.Services.TryAddSingleton<AzureSignalRHostedService>();
-            builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IStartupFilter, AzureSignalRStartupFilter>());
-            builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<MatcherPolicy, NegotiateMatcherPolicy>());
-#endif
-
             return builder;
         }
     }
