+++ src/Microsoft.Azure.SignalR/HubHost/NegotiateHandler.cs
@@ -8,6 +8,7 @@ using System.Security.Claims;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Connections;
 using Microsoft.AspNetCore.SignalR;
+using Microsoft.Azure.SignalR.Common;
 using Microsoft.Extensions.Options;
 
 namespace Microsoft.Azure.SignalR
