+++ src/Microsoft.Azure.SignalR/HubHost/ServiceLifetimeManager.cs
@@ -27,13 +27,10 @@ namespace Microsoft.Azure.SignalR
             ILogger<ServiceLifetimeManager<THub>> logger, AzureSignalRMarkerService marker)
             : base(serviceConnectionManager, protocolResolver)
         {
-            // after core 3.0 UseAzureSignalR() is not required.
-#if !NETCOREAPP3_0
             if (!marker.IsConfigured)
             {
                 throw new InvalidOperationException(MarkerNotConfiguredError);
             }
-#endif
 
             _serviceConnectionManager = serviceConnectionManager;
             _clientConnectionManager = clientConnectionManager;
