+++ src/Microsoft.Azure.SignalR/Utilities/AuthorizeHelper.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authentication;
@@ -16,7 +15,7 @@ namespace Microsoft.Azure.SignalR
     {
         public static async Task<bool> AuthorizeAsync(HttpContext context, IList<IAuthorizeData> policies)
         {
-            if (policies != null && policies.Count == 0)
+            if (policies.Count == 0)
             {
                 return true;
             }
@@ -66,22 +65,6 @@ namespace Microsoft.Azure.SignalR
                 return false;
             }
             return false;
-        }
-
-        public static List<IAuthorizeData> BuildAuthorizePolicy(Type hub)
-        {
-#if NETCOREAPP3_0
-            // Core 3.0 is using AuthorizationMiddleware to handle this, no need to do again under Azure SignalR.
-            return null;
-#else
-            var authorizeAttributes = hub.GetCustomAttributes(typeof(AuthorizeAttribute), inherit: true);
-            var authorizeData = new List<IAuthorizeData>();
-            foreach (var attribute in authorizeAttributes)
-            {
-                authorizeData.Add((AuthorizeAttribute)attribute);
-            }
-            return authorizeData;
-#endif
         }
     }
 }
